+++ frontend/src/contexts/global/global-actions.ts
@@ -287,44 +287,6 @@ export function createActions(
       }
     },
 
-    deactivateWalletAction(wallet: string): GlobalAction {
-      return {
-        type: 'DEACTIVATE_WALLET',
-        wallet
-      }
-    },
-
-    renameWallet(from: string, to: string, onComplete?: () => void) {
-      return async (dispatch: GlobalDispatch) => {
-        logger.debug('RenameWallet')
-
-        try {
-          const passphrase = await requestPassphrase()
-
-          await service.WalletApi.RenameWallet({
-            wallet: from,
-            newName: to,
-            passphrase
-          })
-
-          dispatch({
-            type: 'RENAME_WALLET',
-            from,
-            to
-          })
-
-          onComplete?.()
-        } catch (err) {
-          if (err !== 'dismissed') {
-            AppToaster.show({
-              intent: Intent.DANGER,
-              message: `${err}`
-            })
-          }
-        }
-      }
-    },
-
     // Network actions
 
     changeNetworkAction(network: string) {
